Patch branch relocation with `_unresolved` symbol
When collecting relocation we can detect if it would be a branch relocation and if it's we can resolve it to the `_unresolved` function (if exist) and still write relocation to the module. That way if there was an error during runtime linking, the game would actually call the `_unresolved` function letting you know there was an error.

Signed-off-by: InusualZ <wesmoret98@gmail.com>